cleanup

+++ src/draw/shape.rs
@@ -162,10 +162,10 @@ pub(crate) fn bounding_box_for_points(points: &[(i32, i32)], thick: f64) -> Opti
     if points.is_empty() {
         return None;
     }
-    let mut min_x = points[0].0 as i32;
-    let mut max_x = points[0].0 as i32;
-    let mut min_y = points[0].1 as i32;
-    let mut max_y = points[0].1 as i32;
+    let mut min_x = points[0].0;
+    let mut max_x = points[0].0;
+    let mut min_y = points[0].1;
+    let mut max_y = points[0].1;
 
     for &(x, y) in &points[1..] {
         min_x = min_x.min(x);
